fix typos and rename files properly
