+++ TestResultDBAccess/DBAccess/__init__.py
@@ -1,2 +1,2 @@
-from .direct_db_accesss import DirectDBAccess
+from .direct_db_access import DirectDBAccess
 from .rest_api_db_access import RestApiDBAccess
